improve interface to filter rows
now it's possible to filter some rows in or out on both "plot" and
"diff" commands with an improved interface.
previously, it was possible to exclude cells only for the "plot" command
and to only include cells only for the "diff" command.
arguments "--exclude-label", "--exclude-value", "--include-label" and
"--include-value" are now reduced to only "--filter-label" and
"--filter-value", with the new argument "--filter-mode" to specify
whether to filter in or out the matching rows (same interface for both
commands).